WIP Rename "archive" routes etc. back to "bookmarks"
