+++ app/components/bookmark-item/template.hbs
@@ -17,7 +17,7 @@
     {{/if}}
   </p>
   <div class="actions">
-    {{#link-to "archive.edit" item title=(t "links.edit")}}{{icon-edit}}{{/link-to}}
+    {{#link-to "bookmarks.edit" item title=(t "links.edit")}}{{icon-edit}}{{/link-to}}
   </div>
   {{#if item.description}}<p class="description">{{item.description}}</p>{{/if}}
 </li>
