+++ app/controllers/bookmarks/new.js
@@ -16,7 +16,7 @@ export default Controller.extend({
   bookmarkletUsed: false,
 
   bookmarkletLink: computed(function() {
-    return htmlSafe(`javascript:void(location.href="${window.location.origin}/archive/new/?url="+encodeURIComponent(location.href)+"&title="+encodeURIComponent(document.title))`);
+    return htmlSafe(`javascript:void(location.href="${window.location.origin}/bookmarks/new/?url="+encodeURIComponent(location.href)+"&title="+encodeURIComponent(document.title))`);
   }),
 
   actions: {
