+++ app/templates/bookmarks/index.hbs
@@ -21,7 +21,7 @@
       </div>
       <div class="text">
         <p>{{t "bookmark.index.you_have_not_stored_any_bookmarks_yet"}}</p>
-        <p><LinkTo @route="archive.new" class="button button-secondary">{{t "bookmark.index.add_the_first_one"}}</LinkTo></p>
+        <p><LinkTo @route="bookmarks.new" class="button button-secondary">{{t "bookmark.index.add_the_first_one"}}</LinkTo></p>
         <p class="learn_more"><LinkTo @route="about" class="button button-naked">{{t "bookmark.index.learn_more"}}</LinkTo></p>
       </div>
     </section>
